requirements.txt
nfcpy
pyobject
playsound

